Improve error handling

+++ src/cuckoo/mean_cuckoo.h
@@ -14,18 +14,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * In addition, as a special exception, the copyright holders give 
- * permission to link the code of portions of this program with the 
- * Botan library under certain conditions as described in each 
- * individual source file, and distribute linked combinations 
+ * In addition, as a special exception, the copyright holders give
+ * permission to link the code of portions of this program with the
+ * Botan library under certain conditions as described in each
+ * individual source file, and distribute linked combinations
  * including the two.
  *
- * You must obey the GNU General Public License in all respects for 
- * all of the code used other than Botan. If you modify file(s) with 
- * this exception, you may extend this exception to your version of the 
- * file(s), but you are not obligated to do so. If you do not wish to do 
- * so, delete this exception statement from your version. If you delete 
- * this exception statement from all source files in the program, then 
+ * You must obey the GNU General Public License in all respects for
+ * all of the code used other than Botan. If you modify file(s) with
+ * this exception, you may extend this exception to your version of the
+ * file(s), but you are not obligated to do so. If you do not wish to do
+ * so, delete this exception statement from your version. If you delete
+ * this exception statement from all source files in the program, then
  * also delete it here.
  */
 #ifndef MERIT_CUCKOO_MEAN_CUCKOO_H
@@ -43,13 +43,13 @@ namespace merit
         // Find proofsize-length cuckoo cycle in random graph
         bool FindCycle(
                 const char* hex_header_hash,
-                size_t hex_header_hash_len,
+                uint32_t hex_header_hash_len,
                 uint8_t edgeBits,
                 uint8_t proofSize,
                 std::set<uint32_t>& cycle,
                 size_t threads_number,
                 ctpl::thread_pool&);
-    } 
+    }
 }
 
 #endif // MERIT_CUCKOO_MEAN_CUCKOO_H
